backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
python-dotenv==1.0.1